requirements.txt
asgiref==3.7.2
astroid==2.15.6
autopep8==2.0.2
beautifulsoup4==4.12.2
dill==0.3.7
Django==4.1
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-js-asset==2.1.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysqlclient==2.2.0
Pillow==10.0.0
platformdirs==3.10.0
pycodestyle==2.11.0
pylint==2.17.5
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
wrapt==1.15.0
